HEALIX_HMS/healix-hms/services/patient_service/requirements.txt
fastapi
uvicorn
motor
python-dotenv
pydantic
pydantic-settings
email-validator
httpx
httpx
